+++ src/pages/Login/Login.tsx
@@ -1,9 +1,7 @@
-
-import { useContext, useEffect } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
-import { RouterPath } from "@/routes/path";
+import { useContext } from "react";
 import { AuthContext } from "@/context/AuthContext";
 import logo from "@/assets/images/logo.png";
+import { useAuthRedirect } from "@/hooks/useAuthRedirect";
 import NavigationBar from "@/components/NavigationBar/NavigationBar";
 import {
   Wrapper,
@@ -17,8 +15,7 @@ import { useLoginFormValidation } from "@/hooks/useLoginFormValidation";
 
 const LoginPage = () => {
   const auth = useContext(AuthContext);
-  const navigate = useNavigate();
-  const location = useLocation();
+  useAuthRedirect();
   const {
     email,
     password,
@@ -31,19 +28,10 @@ const LoginPage = () => {
     isValid,
   } = useLoginFormValidation();
 
-  useEffect(() => {
-    if (auth?.user) {
-      const redirectTo = location.state?.from || RouterPath.MYPAGE;
-      navigate(redirectTo, { replace: true });
-    }
-  }, [auth, navigate, location.state]);
-
   const handleLoginClick = (e: React.FormEvent) => {
     e.preventDefault();
     if (isValid && auth) {
       auth.login({ email });
-      const redirectTo = location.state?.from || RouterPath.MYPAGE;
-      navigate(redirectTo, { replace: true });
     }
   };
